+++ storage/innobase/srv/srv0srv.cc
@@ -192,7 +192,7 @@ const ulint	srv_buf_pool_min_size	= 5 * 1024 * 1024;
 /** Default pool size in bytes */
 const ulint	srv_buf_pool_def_size	= 128 * 1024 * 1024;
 /** Requested buffer pool chunk size */
-ulong	srv_buf_pool_chunk_unit;
+size_t	srv_buf_pool_chunk_unit;
 /** innodb_lru_scan_depth; number of blocks scanned in LRU flush batch */
 ulong	srv_LRU_scan_depth;
 /** innodb_flush_neighbors; whether or not to flush neighbors of a block */
